Add grpc tests
